[Tokyo Xtreme Racer 3] - Pnach file for the patched version (#294)
* [Tokyo Xtreme Racer 3] - Pnach file for the patched iso of the game

When Tokyo Xtreme Racer 3 released in America, the localization team made one big mistake. They converted all the Credits from Japanese Yen to American Dollars by dividing the value by 100. You start with 40,000 credits instead of 4,000,000 million credits and so on.

recently a patch has been created in other to apply it to game so that no cheat codes are needed anymore

https://www.romhacking.net/hacks/8339/

However widescreen and No-Interlacing patches are missing for the patched iso

This PR will address that

* fix naming convention