GTA San Andreas Widescreen improvement and Silentpatch Lite for 399A49CA and 2C6BE434
The current widescreen patch, while functional, has the HUD and the radar unaffected. I include pnach files for 1.03 and 3.00 that make the widescreen appear more like ThirteenAG's widescreen mod on PC, thus fixing the HUD and radar. It also has the added benefit of not glitching the emergency vehicles. I also include a Silentpatch lite that adds cosmetic and QOL enhancements, similar to the PC version, while leaving game mechanics otherwise untouched, so it wouldn't trigger RetroAchievements' anticheat system.

1.03
![Grand Theft Auto - San Andreas_SLUS-20946_20250108112003](https://github.com/user-attachments/assets/075c60f5-614b-49fc-b706-a8917196c333)

3.00
![Grand Theft Auto - San Andreas_SLUS-20946_20250108130059](https://github.com/user-attachments/assets/257ef544-e877-4557-b52c-d63861b027ce)

(For best results, turn the ingame widescreen option on, but that does not affect the radar/HUD whether on or off)