NFS Carbon/ProStreet: Correct the serial the game uses to detect Most Wanted Black Edition saves (#499)
**Please merge together with
https://github.com/PCSX2/pcsx2/pull/12340.**

Fixes an old bug where only base Most Wanted saves granted the cash
bonus, and MW Black Edition saves did not. This is because the game used
an incorrect serial to detect BE saves.


![image](https://github.com/user-attachments/assets/d1bbaaba-c283-4518-8af6-6e139dc24460)

ProStreet is seemingly unaffected, as it detects Most Wanted, Most
Wanted BE, Carbon, Carbon CE correctly.