patches: NFS Underground: SilentPatch (#519)
Featured fixes:

* Fixed the unlock requirements for Magazine 22, which is awarded for
setting a Drift Track record. The magazine is now awarded for beating
the average lap high score, ensuring the requirements are fair
regardless of the number of laps in the event.
* Drift events now display the total score as their High Score instead
of style points.
* Fixed a buffer overflow issue in the UI display of opponent names.
This resolves a bug where Samantha's name appeared corrupted during
races.

Same as in
https://cookieplmonster.github.io/mods/need-for-speed-underground-ps2/#silentpatch,
submitted to the patch db so it can be used with RetroAchievements
Hardcore Mode.