patches: Add 60fps 007 Quantum of Solace US patch
This patch adds 60fps functionality to the US version of 007 - Quantum of Solace. The patch conditionally switches to 30fps when .pss files run, to prevent them running too quickly. It also affects the weapon recoil compensation, which seems to be affected by framerate, to recreate similar recoil behavior as if the game was running at 30fps.