patches: Sega Rally 2006, Tokyo Xtreme Racer Drift 2 new patches. (#536)
Added 1 patch by JakeMR2 (posted in Reddit some time ago, and properly
tested without any problem) for Tokyo Xtreme Racer 2. The other ones are
by Racer_2 published this time on YouTube (there is a thread on PCSX2
forums but it basically contains the same patches) for Sega Rally 2006.
I also made reformatted the widescreen patch for the latter game in
order to comply with pnach 2.0 standards.