Ultimate Spider-man update patches (#558)
Updated patches. I found the rendering solution for widescreen a while
ago. The only thing missing are the flashes that appear in the game when
an enemy fires lightning bolts or in the storm where we fight Venom at
the beginning and end of the game. These flashes are in 4:3. I don't
think they're very relevant, and all things considered, the positives
outweigh the negatives.
I think it's a good time to add it to the repository so we can enjoy it.

Fixes #556 

![Ultimate
Spider-Man_SLES-53391_20250512112758](https://github.com/user-attachments/assets/0fa7815d-f2d0-417e-bd46-935ded3e147f)
![Ultimate
Spider-Man_SLES-53391_20250512112809](https://github.com/user-attachments/assets/c028f560-2cee-4ebb-afbe-e110d59eff77)