ws: Alien Hominid: Disable the widescreen fix breaking respawning (#579)
Levels like 1-2 have scrolling sections where the player jumps over
vehicles and they can die if they land on the road. With the current
widescreen fix enabled, if the player dies outside of the 4:3 screen
area, the game softlocks and they won't respawn.


This bug renders the game essentially unplayable.