Updates Kingdom Hearts Final Mix I & II  (#597)
Hey!

I updated the patches for Kingdom Hearts Final Mix I & II.
Basically I:
- Changed the name for the new update of the english patch of KHFM1 (CRC
is now BD3FB870 as stated in #565)
- Uncommented the 60FPS patch for KHFM1 (english patch)
- Uncommented the 60FPS patch for KHFM2 (english patch)
- Added file SLPM-66675_E6FB8E10.pnach which is the english patch
provided by RetroAchievments (see
https://retroachievements.org/game/2781/hashes), which is the same as
the one created with Crazycatz00 patcher, but it uses a xdelta file
instead (same patches used).
- Globally added some markers to the patches for selection

I think uncommenting the 60 FPS patches would be a good idea because
using RetroAchievments in Hardcore Mode only allows for patches, not
cheats. Adding [60 FPS] allows one to select it or not if there is some
issues with it (but I just played through KHFM1 with 60FPS - worked well
- and quickly tested it for KHFM2 and seemed to work).
Didn't touch "Disable image map names" in KHFM2 because not sure what it
is doing.