Rugrats: Royal Ransom (US): Fix horz. camera speed at 60fps, add uninverted camera patch (#622)
The horizontal movement speed of the third-person camera is twice as
slow at 60fps, due to the delta-time value accidentally being applied to
the horizontal movement delta twice. This pull request fixes that issue.
I've also added an uninverted camera patch. While the game provides a
means to uninvert the horizontal axis, the vertical axis is stuck as
inverted by default.