Ys: The Ark of Napishtim (US) - Fix analog movement deadzone in top left (#631)
When moving the player character with the analog stick, the game will
explicitly skip movement processing logic if the both the analog axes X
and Y are equal to 0. Likely, the programmer thought that this meant
that the analog stick was neutral (as opposed to the reality of it being
in the top left).

This patch removes those checks. I haven't extensively tested this
patch, but I can't imagine it causing any issues -- the player is now
able to move when you move the analog stick exactly to the top left.