Test Drive Unlimited: Trigger controls mapping (#634)
A cheat code re-enabling a previously unused control mapping in Test
Drive Unlimited, and swapping L1/R1 buttons for L2/R2 triggers. With
this change, this mapping matches “modern” controls of the X360 and PC
versions of Test Drive Unlimited.

Also added support for `SLES-54466`, as it has the same executable as
`SLES-53753`, just a different serial.