Fatal Zero hack disabling (#637)
Disables the Cutscene Render fix for the NTSC-J & NTSC-K versions of the
disc.
The NTSC-U disc does not have this hack.

With the PAL disc, this hack wracks some rendering of the game.
After testing, it turned out that the hack wasn't needed anymore anyway,
 and was thus disabled. See more here: #591.

It's resonable to assume that the hack causes the same wracking
 with the NTSC-J & NTSC-K discs and isn't needed anymore anyway
 (as with the PAL disc), however testing is appreciated.