Ultimate Spider-Man update (#641)
In this issue: https://github.com/PCSX2/pcsx2/issues/13679 
It's been reported that there's a problem with the FPS patch, but my
research suggests the patch works perfectly (I only fixed the
conditional for sync in FMV).
The problem is caused by widescreen, which definitely needs to be
disabled since no better solution has been found.

The video shows that the car looks good and you can have a good time
with this FPS patch enabled and without the widescreen patch.


https://github.com/user-attachments/assets/f4955a5a-7f8a-46ea-89ae-7b39fcef52c2